Extract ApplicationForm component
